task: getting rid off nividia (at least for now)
In the past I had often short freezes with my keyboard and mouse. These
seem to be connected with the nvidia driver. After a freeze the was
allways the message (journalctl --follow):
kernel: pcieport 0000:00:01.0: PME: Spurious native interrupt!
Since I disabled the nvidia driver it is gone (and so are the freezes).
Maybe someday I will give a new try with the open source driver.